requirements.txt
numpy
pyaudio
pyrtlsdr
tkinter
ttkthemes